+++ bw.workflow.fingerprintsinquiry/src/main/java/sa/gov/nic/bio/bw/workflow/fingerprintsinquiry/gui/RectangleWithLabel.java
@@ -1,8 +1,0 @@
-package sa.gov.nic.bio.bw.workflow.fingerprintsinquiry.gui;
-
-import java.awt.Rectangle;
-
-public class RectangleWithLabel extends Rectangle
-{
-
-}