+++ RELEASE-NOTES.md
@@ -14,6 +14,8 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 ## v2019.01.1
 
 - NEW: Add Delete-Convicted-Report workflow.
+- NEW: Add WEBSERVICES-LIST.md.
+- CHN: Change the workflow fingerprintsinquiry to civilcriminalfingerprintsinquiry.
 
 ---